+++ src/Services/FeatureLayer.js
@@ -1,17 +1,17 @@
-L.esri.Services.FeatureLayer = L.esri.Service.extend({
+L.esri.Services.FeatureLayer = L.esri.Services.Service.extend({
 
   query: function(){
-    return new L.esri.Services.Query(this);
+    return new L.esri.Tasks.Query(this);
   },
 
   createFeature: function(feature, callback, context) {
     feature = L.esri.Util.geojsonToArcGIS(feature);
-    this.post(this.url + 'addFeatures', JSON.stringify(feature), callback, context);
+    this.post(this.url + 'addFeatures', JSON.stringify([feature]), callback, context);
   },
 
   updateFeature: function(feature, callback, context) {
     feature = L.esri.Util.geojsonToArcGIS(feature);
-    this.post(this.url + 'updateFeatures', JSON.stringify(feature), callback, context);
+    this.post(this.url + 'updateFeatures', JSON.stringify([feature]), callback, context);
   },
 
   deleteFeature: function(id, callback, context) {